Move AbstractGeneratorTest to `generators-test`
