+++ core/src/commonTest/kotlin/com/github/jcornaz/kwik/generator/LongGeneratorTest.kt
@@ -1,6 +1,6 @@
 package com.github.jcornaz.kwik.generator
 
-import com.github.jcornaz.kwik.AbstractGeneratorTest
+import com.github.jcornaz.kwik.generator.test.AbstractGeneratorTest
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
